Refactor integration test helpers to separate project
Upgraded some nuget packages as part of refactor
