+++ Portal.Tests/Integration/BasicAccessTests.cs
@@ -2,8 +2,11 @@
 using System.Net.Http;
 using System.Threading.Tasks;
 using FluentAssertions;
+using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Mvc.Testing;
+using Microsoft.Extensions.DependencyInjection;
 using Portal.Tests.Integration.Infrastructure;
+using Test.Helpers.Integration;
 using Xunit;
 
 namespace Portal.Tests.Integration
@@ -18,6 +21,12 @@ namespace Portal.Tests.Integration
         {
             httpClient = factory
                 .WithConnectionString(dbFixture.ConnectionString)
+                .WithTestServices(services =>
+                {
+                    services.AddAuthentication("Test")
+                        .AddScheme<AuthenticationSchemeOptions, TestAuthHandler>(
+                            "Test", _ => { });
+                })
                 .CreateClient(new WebApplicationFactoryClientOptions
                 {
                     AllowAutoRedirect = false
