+++ Portal.Tests/Integration/Infrastructure/DatabaseCollection.cs
@@ -1,0 +1,11 @@
+﻿using Test.Helpers.Integration;
+using Xunit;
+
+namespace Portal.Tests.Integration.Infrastructure
+{
+    [CollectionDefinition(CollectionName)]
+    public class DatabaseCollection : ICollectionFixture<DlcsDatabaseFixture>
+    {
+        public const string CollectionName = "Database Collection";
+    }
+}
