+++ Portal.Tests/Integration/Infrastructure/LocalStackCollection.cs
@@ -1,0 +1,11 @@
+﻿using Test.Helpers.Integration;
+using Xunit;
+
+namespace Portal.Tests.Integration.Infrastructure
+{
+    [CollectionDefinition(CollectionName)]
+    public class LocalStackCollection : ICollectionFixture<LocalStackFixture>
+    {
+        public const string CollectionName = "LocalStack Collection";
+    }
+}
