+++ Portal.Tests/Integration/SpacesTests.cs
@@ -6,8 +6,11 @@ using AngleSharp.Html.Parser;
 using DLCS.Repository;
 using DLCS.Repository.Entities;
 using FluentAssertions;
+using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Mvc.Testing;
+using Microsoft.Extensions.DependencyInjection;
 using Portal.Tests.Integration.Infrastructure;
+using Test.Helpers.Integration;
 using Xunit;
 
 namespace Portal.Tests.Integration
@@ -24,6 +27,12 @@ namespace Portal.Tests.Integration
             dbContext = dbFixture.DbContext;
             httpClient = factory
                 .WithConnectionString(dbFixture.ConnectionString)
+                .WithTestServices(services =>
+                {
+                    services.AddAuthentication("Test")
+                        .AddScheme<AuthenticationSchemeOptions, TestAuthHandler>(
+                            "Test", _ => { });
+                })
                 .CreateClient(new WebApplicationFactoryClientOptions
                 {
                     AllowAutoRedirect = false
