+++ protagonist.sln
@@ -43,6 +43,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Applications", "Application
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Orchestrator.Tests", "Orchestrator.Tests\Orchestrator.Tests.csproj", "{7A0F20A1-AED0-4771-A90D-C150BAFED464}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test.Helpers", "Test.Helpers\Test.Helpers.csproj", "{68876B84-0636-4573-A53A-926AB5C16231}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -121,6 +123,10 @@ Global
 		{7A0F20A1-AED0-4771-A90D-C150BAFED464}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7A0F20A1-AED0-4771-A90D-C150BAFED464}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7A0F20A1-AED0-4771-A90D-C150BAFED464}.Release|Any CPU.Build.0 = Release|Any CPU
+		{68876B84-0636-4573-A53A-926AB5C16231}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{68876B84-0636-4573-A53A-926AB5C16231}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{68876B84-0636-4573-A53A-926AB5C16231}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{68876B84-0636-4573-A53A-926AB5C16231}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -141,5 +147,6 @@ Global
 		{D8A867D1-0DD3-4532-A12D-04529A4FCD3F} = {820EE99A-E129-44EF-A7BD-519BCE6167D3}
 		{207C1A36-B240-457B-AEF9-EF6836BDF58A} = {820EE99A-E129-44EF-A7BD-519BCE6167D3}
 		{7A0F20A1-AED0-4771-A90D-C150BAFED464} = {89603DC1-C3AD-49C8-A412-D556AE6483D6}
+		{68876B84-0636-4573-A53A-926AB5C16231} = {89603DC1-C3AD-49C8-A412-D556AE6483D6}
 	EndGlobalSection
 EndGlobal
